+++ R/db.R
@@ -143,6 +143,8 @@ emr_db.connect <- function(db_dirs = NULL, load_on_demand = NULL, do_reload = FA
             }
         }
     )
+
+    purrr::walk(db_dirs, emr_entries.reload)
 }
 
 #' @export emr_db.init
@@ -240,6 +242,8 @@ emr_db.reload <- function() {
             }
         }
     )
+
+    purrr::walk(emr_db.ls(), emr_entries.reload)
 }
 
 #' Unload all tracks from naryn database