+++ R/zzz.R
@@ -4,6 +4,7 @@
 #' @export
 #' @noRd
 .naryn <- new.env(parent = emptyenv())
+.naryn$entries <- list()
 
 .onLoad <- function(lib, pkg) {
 }